// generated/src/app/api-models/site-address-contact.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class SiteAddressContact {
  siteCode: string;
  seqSiteAddressId: number;
  seqSiteContactId: number;
  contactTitle: string;
  contactName: string;
  phoneNumber: string;
  extension: string;
  faxNumber: string;
  primaryContact: string;
  primaryDistributionMethod: string;
  emailId: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateUser: string;
  updateProcess: string;
  updateDatetime: Date;
}